fix: moved environment configuration inside the templated ssr
