+++ packages/template-cli/templates/src/ssr/environment-configuration/environment-variables.js
