+++ packages/template-cli/templates/src/ssr/environment-configuration/index.js
